requirements.txt
google-cloud-storage>=1.42.3
google-cloud-datacatalog>=3.4.3
google-cloud-bigquery>=2.28.1
google-cloud-dlp>=1.1.0